rename: 예시용 테이블의 폴더 구조 변경
