+++ src/components/common/Table/Assemble/UnRegisteredDataTable.tsx
@@ -1,10 +1,10 @@
 import { IUnRegisteredData } from "@/types/dto.temp";
 
-import TableBodyUI from "./Composition/TableBodyUI";
-import TableCellUI from "./Composition/TableCellUI";
-import TableContainerUI from "./Composition/TableContainerUI";
-import TableHeaderUI from "./Composition/TableHeaderUI";
-import TableRowUI from "./Composition/TableRowUI";
+import TableBodyUI from "../Composition/TableBodyUI";
+import TableCellUI from "../Composition/TableCellUI";
+import TableContainerUI from "../Composition/TableContainerUI";
+import TableHeaderUI from "../Composition/TableHeaderUI";
+import TableRowUI from "../Composition/TableRowUI";
 
 interface UnRegisteredDataTableProps {
   data: IUnRegisteredData[]
